EC-14148. Обернул получение сохраненного typingIntervalSeconds корутинкой. Плюс мелкие правки.

+++ src/main/java/im/threads/business/rest/models/ConfigResponse.kt
@@ -8,5 +8,5 @@ data class ConfigResponse(
 )
 
 data class Settings(
-    var typingMessagesIntervalSeconds: Long = 3
+    var typingMessagesIntervalSeconds: Int = 3
 )
